backend/requirements.txt
fastapi[all]
langgraph
langchain
langchain-google-genai
langchain-openai
pydantic-settings
python-dotenv
arxiv
wikipedia
pydantic
langchain-groq
langchain_google_genai
uvicorn
langchain_community 
wikipedia
langchain_core
langchain-ollama
tavily-python
langchain_tavily
opik
mem0ai
langfuse
